requirements.txt
discord.py
openai
python-dotenv
flask
PyMuPDF